week1_2020101104.md
## WEEK 1,LECTURE 1


DATE:18/08/2021

This was the first class of algorithms and analysis class.

Sir has taught the Keywords in this class they are,

1.Compuational problems
2.Computational Solutions
3.Best solution for Computational problem
4.Problems as memebership queries in languages
5.Axioms of computation

------> BEGINNING <---------

I Have understood that there are many ways of posting problems ,a single problem can post in many ways.

There are many problems in this world but we are not going to talk about all those problems. We will talk about the computational problems only.we are not going to consider general problems such as outside and personal problems.


----> COMPUTATIONAL PROBLEMS <------

* Suppose if we have a problem to sort the numbers:

Mythological solution :We should meditate like loop till god appears,after god appears we will give the problem list to god
Now god will sort the numbers and return as the list of numbers in the sorted order.

Next Generation solution:

first we should get married and have children and put them in Good university or college and after he learns sorting algorithm you can give you problem to him and he will sort the numbers and give you numbers.

But we don't want these types of solutions we want computational solutions to the computational problems.

 
-----> DEFINITION <------

Computational Problems:
whether we can post the membership query in the language or not.We wont consider afghanisthan problems as computational because we can not post them as membership query in language.


-----> COMPUTATIONAL SOLUTIONS <------


Computational Solutions:

If they satisfy the three axioms of computation. we can conclude them as computaional solutions.

they are,
* Machines are Not omnipresent
-> That means it takes non-zero time to retrive the infromation from faroff locations.

* Machines are not omniscient
-> finite amount of information is stored/retrived from the finite volume


* Machines are not omnipotent
-> finite length code gives finite amount of control



-----> WHICH IS BEST <--------


After declaring them as computational problem and computational solutions we will now look in to whether they are best solution or not.

if not, which is the best one in,

* Space
* Time
* Worst case analysis

and what is the use of taking the best solution for use....all these sir taught us in the class.
and it is more important to that because Best solutions works and far better than any other solutions


--------> METHODS OF LEARNING <--------

Sir has also said about the methods of learning ,They are

* Self Discovery

We should understand the problem and we should think over it how to solve that ,We should self discover the problem and should try to 
solve the problem in a optimistic way

* reading and understanding

To solve the problem First main important thing is we should read and understand it properly so we should read and understand properly

* Brain Storming in Teams

It is a method of generating ideas and sharing knowledge to solve a particular problem,

------->ENDING<--------


And In class sir has also taught that we are going to learn new techniques to solve the problems ,they are

* DP
* Divide and Conquere
* Linear programming
* Greedy

These are techniques we implement to find the better and better solutions, till we hit the optimistic solution.


## WEEK_1, LECTURE 2 

## WEEK 1,Lecture 2

DATE:21/08/2021
Saturday

This was the Second class of algorithms and analysis class.

-------> RECALL OF PREVIOUS CLASS <-------


Sir had small recap of the previous class,they are

* what are computational problems
* what are computational solutions
* What are Best and optimistic solutions
* Methods of learning
* Problems in membership queries


-------> COMPUTATIONAL PROBLEMS THAT COMPUTER CANNOT SOLVE <--------

* There are many computational problems which computers cannot solve
* There are uncountably infinite problems to solve
* But there are computer programmes only which are countable
* there is a proof for this


------> COUNTABILITY <--------

* An Infinite set is countable if there exists a bijection between natural numbers and this set
* Sir has also discussed about the theorem "The set of finite length bit strings is countable i.e {0,1}*.

* We can make one-to-one correspondence from Natural numbers to set of finite length bit strings

1 -> e
2 -> 0
3 -> 1
4 -> 00
5 -> 01
6 -> 10
7 -> 11
8 -> 000
9 -> 001
10 ->010

and so on...
 
-----> UNCOUNTABILITY <-------

* Set is uncountable if it is not countable.
* If we are able to prove that power set of {0,1}* is uncountable then
Then we are done with proving that they are uncountably computational problems that computers can not solve.

By cantor's Diabolical diagonalization proves that Power set of {0,1}* is not countable



----> Cardinality(problems)==Cardinality({0,1}*)<--------

* Every computer programme can be encoded as binary string.
* So the cardinality of the computer programmes is no more than {0,1}*
* So we can confirm that number of computer programmes are countable
* But number of problems is uncountable( as problems are membership queries in languages)
* So it has the same cardinality as power set of {0,1}*


i.e ------> cardinality(problems)==cardinality(P({0,1})*)

Hence it is proved that There are problems than the computer problems


# THERE are uncountably infinite problems than the computer programmes and also there are infinitely many problems that does not have a computer programme.


* In future the problems for which computer solutions does not exist now...they will come to existence in future.


Sir has also talked again about he attributes of the machines.


------> ATTRIBUTES OF MACHINES <-----------

* Machines are not omnipresent

  It will take the non-zero time to retrieve the data from far-off locations.

* Also they are not omniscient

  Finite information can be stored/retrieved from finite volume

* Also they are not omnipotent

  Finite length of code only exerts finite amount of control.


* Sir has also introduced about the turing machines 

-----> TURING MACHINES <-------

* Turing machine is a mathematical model which consists of infinite length tape divided into cells

* It handles a string of 0's and 1's.